backend/requirements.txt
fastapi
uvicorn
duckdb
pandas
pandas_ta
python-multipart
# dev dependencies
pytest
httpx
smartapi-python
pyotp
python-dotenv
